Update config.php
Sorry for messing around, is my first commit and had some issues doing it (not sure if is done properly, though).
So the main change here is I added back the "$url" that was missing after the last update in Jan 2019 and added the new API address from toggl as they changed the API at the end of June 2021.
This makes the system work again. 

Many thanks for this contribution. It has been really helpful to me for the last few years.